Persist MQTT payload and update schemas/models

+++ models/study_log.py
@@ -10,7 +10,7 @@ class Studylog(Base):
     date_time = Column(DateTime, nullable=False)
     sound_score = Column(Integer, nullable=False)
     light_score = Column(Integer, nullable=False)
-    temperature_score = Column(Integer, nullable=False)
+    temperature_score = Column(Float, nullable=False)
     focus_score = Column(Integer, nullable=False)
     score = Column(Integer, nullable=False) 
     is_uncomfortable = Column(Boolean, nullable=False)
